+++ app/src/main/java/com/example/foosball/models/FoosmenBar.java
@@ -1,5 +1,0 @@
-package com.example.foosball.models;
-
-public class FoosmenBar {
-
-}